move twetterService into Usecase
