+++ handler/router.go
@@ -2,18 +2,18 @@ package handler
 
 import (
 	"github.com/a-omori-yumemi/YumetterAPI/repository"
-	"github.com/a-omori-yumemi/YumetterAPI/service"
+	"github.com/a-omori-yumemi/YumetterAPI/usecase"
 	"github.com/labstack/echo/v4"
 )
 
-func SetRoute(e *echo.Echo, repos repository.Repositories, services service.Services) {
+func SetRoute(e *echo.Echo, repos repository.Repositories, usecases usecase.Usecases) {
 	g := e.Group("/v1")
 
 	tweetsg := g.Group("/tweets")
-	tweetsg.GET("/:tw_id", GetTweet(services.TweetService))
-	tweetsg.DELETE("/:tw_id", DeleteTweet(services.TweetService))
-	tweetsg.GET("/", GetTweets(services.TweetService))
-	tweetsg.POST("/", PostTweet(services.TweetService))
+	tweetsg.GET("/:tw_id", GetTweet(usecases.TweetService))
+	tweetsg.DELETE("/:tw_id", DeleteTweet(usecases.TweetService))
+	tweetsg.GET("/", GetTweets(usecases.TweetService))
+	tweetsg.POST("/", PostTweet(usecases.TweetService))
 
 	usersg := g.Group("/users")
 	usersg.GET("/:usr_id", GetUser(repos.UserRepo))
