--- service/service.go
@@ -1,5 +1,0 @@
-package service
-
-type Services struct {
-	TweetService ITweetService
-}
