+++ usecase/tweet_impl.go
@@ -1,4 +1,4 @@
-package service
+package usecase
 
 import (
 	"database/sql"
