+++ usecase/usecase.go
@@ -2,4 +2,5 @@ package usecase
 
 type Usecases struct {
 	Authenticator IAuthenticator
+	TweetService  ITweetService
 }
